drivers: serial: nrfx_uarte: Increase hidden timeouts
In uarte_nrfx_poll_out, a previous commit set a short
timeout on sending a character of 1ms. This is
insufficient when very low baud rates and/or
hardware flow control is enabled.

Ideally, the fix would be to either add a timeout
parameter and a return code from the function, or,
a Kconfig parameter which allows for infinite timeouts,
rather than a hard-coded amount.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>