shell: debug improvements
Improve display of threads (include entry point).

Add ability to stop argument filtering after a certain
argument index. This is needed to allow nrf9160 AT
commands containing certificates to pass through the
shell unmolested.

Add flag to shell_flags to control echo as obscure
character (for passwords) or not. Default to obscure
(need to add Kconfig flag for this).

Handle log message queueing error better.

Normally, if a shell command is compiled out using
SHELL_COND_CMD(), a line for this command will still
be printed but will be blank. Change it so compiled
out commands are not listed as blank lines.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>