net: sockets: tls: Implement TLS_DTLS_CONNECTION_ID option
Implement TLS_DTLS_CONNECTION_ID socket option, which enables to use
Connection ID extension for the DTLS session.

The option allows to set the value and the length of the CID to use with
`setsockopt()` function. Setting the CID length to 0, enables the
extension but does not send the own CID to the peer, as described in the
specification.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>